# HUCE-ISRS-BE/university_system/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "University System API",
        "description": "API hệ thống quản lý trường đại học – dùng để tích hợp với hệ thống đăng ký học phần bổ sung",
        "contact": {
            "email": "admin@university.edu.vn"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8001",
            "description": "University System Server"
        }
    ],
    "paths": {
        "/api/token": {
            "post": {
                "tags": [
                    "Xác thực"
                ],
                "summary": "Lấy access token",
                "description": "Xác thực hệ thống client bằng client_id và client_secret. Trả về access_token để sử dụng cho các API tiếp theo (machine-to-machine authentication).",
                "operationId": "getToken",
                "requestBody": {
                    "description": "Thông tin xác thực client",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "client_id",
                                    "client_secret"
                                ],
                                "properties": {
                                    "client_id": {
                                        "description": "Mã định danh của hệ thống client",
                                        "type": "string",
                                        "example": "remedial_system"
                                    },
                                    "client_secret": {
                                        "description": "Mật khẩu bí mật của client",
                                        "type": "string",
                                        "example": "secret_key_here"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Lấy token thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Xác thực thành công"
                                        },
                                        "data": {
                                            "properties": {
                                                "access_token": {
                                                    "type": "string",
                                                    "example": "abc123..."
                                                },
                                                "token_type": {
                                                    "type": "string",
                                                    "example": "Bearer"
                                                },
                                                "expires_in": {
                                                    "type": "integer",
                                                    "example": 3600
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "errors": {
                                            "type": "null"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Xác thực thất bại – client_id hoặc client_secret không hợp lệ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Xác thực thất bại"
                                        },
                                        "data": {
                                            "type": "null"
                                        },
                                        "errors": {
                                            "type": "null"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Dữ liệu đầu vào không hợp lệ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Dữ liệu không hợp lệ"
                                        },
                                        "data": {
                                            "type": "null"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/student/login": {
            "post": {
                "tags": [
                    "Xác thực"
                ],
                "summary": "Đăng nhập sinh viên (Backdoor)",
                "description": "Đăng nhập dành cho sinh viên sử dụng mã sinh viên và mật khẩu backdoor hệ thống.",
                "operationId": "studentLogin",
                "requestBody": {
                    "description": "Thông tin đăng nhập sinh viên",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "student_id",
                                    "password"
                                ],
                                "properties": {
                                    "student_id": {
                                        "description": "Mã sinh viên",
                                        "type": "string",
                                        "example": "SV001"
                                    },
                                    "password": {
                                        "description": "Mật khẩu backdoor",
                                        "type": "string",
                                        "example": "nuce_backdoor_2026"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Đăng nhập thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Đăng nhập thành công"
                                        },
                                        "data": {
                                            "properties": {
                                                "access_token": {
                                                    "type": "string",
                                                    "example": "abc123..."
                                                },
                                                "token_type": {
                                                    "type": "string",
                                                    "example": "Bearer"
                                                },
                                                "expires_in": {
                                                    "type": "integer",
                                                    "example": 3600
                                                },
                                                "student": {
                                                    "$ref": "#/components/schemas/StudentDto"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "errors": {
                                            "type": "null"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Sai mã sinh viên hoặc mật khẩu backdoor"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StudentDto": {
                "description": "Thông tin sinh viên",
                "properties": {
                    "id": {
                        "description": "Mã sinh viên",
                        "type": "string",
                        "example": "SV001"
                    },
                    "fullName": {
                        "description": "Họ và tên đầy đủ",
                        "type": "string",
                        "example": "Nguyễn Văn An"
                    },
                    "gender": {
                        "description": "Giới tính",
                        "type": "string",
                        "example": "Nam"
                    },
                    "dateOfBirth": {
                        "description": "Ngày sinh",
                        "type": "string",
                        "example": "2002-05-15"
                    },
                    "placeOfBirth": {
                        "description": "Nơi sinh",
                        "type": "string",
                        "example": "Hà Nội"
                    },
                    "personalEmail": {
                        "description": "Email cá nhân",
                        "type": "string",
                        "example": "an@gmail.com"
                    },
                    "universityEmail": {
                        "description": "Email trường",
                        "type": "string",
                        "example": "sv001@university.edu.vn",
                        "nullable": true
                    },
                    "gpaScale10": {
                        "description": "Điểm TB tích lũy hệ 10",
                        "type": "number",
                        "format": "float",
                        "example": 7.5,
                        "nullable": true
                    },
                    "gpaScale4": {
                        "description": "Điểm TB tích lũy hệ 4",
                        "type": "number",
                        "format": "float",
                        "example": 2.8,
                        "nullable": true
                    },
                    "gradeClassification": {
                        "description": "Xếp loại học lực",
                        "type": "string",
                        "example": "Khá",
                        "nullable": true
                    },
                    "totalCredits": {
                        "description": "Tổng tín chỉ tích lũy",
                        "type": "integer",
                        "example": 90,
                        "nullable": true
                    },
                    "failedCredits": {
                        "description": "Số tín chỉ chưa đạt",
                        "type": "integer",
                        "example": 3,
                        "nullable": true
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Nhập access_token nhận được từ POST /api/token",
                "bearerFormat": "token",
                "scheme": "bearer"
            }
        }
    },
...
}